+++ src/providers/auth.ts
@@ -21,42 +21,70 @@
 // SOFTWARE.
 
 import {
-  UserRecord,
-  UserInfo,
-  UserRecordMetadata,
-  userRecordConstructor,
-} from '../common/providers/identity';
-import {
+  BlockingFunction,
   CloudFunction,
   Event,
   EventContext,
   makeCloudFunction,
+  optionsToEndpoint,
+  optionsToTrigger,
 } from '../cloud-functions';
+import {
+  AuthBlockingEventType,
+  AuthEventContext,
+  AuthUserRecord,
+  BeforeCreateResponse,
+  BeforeSignInResponse,
+  HttpsError,
+  UserInfo,
+  UserRecord,
+  userRecordConstructor,
+  UserRecordMetadata,
+  wrapHandler,
+} from '../common/providers/identity';
 import { DeploymentOptions } from '../function-configuration';
 
 // TODO: yank in next breaking change release
 export { UserRecord, UserInfo, UserRecordMetadata, userRecordConstructor };
+
+export { HttpsError };
 
 /** @hidden */
 export const provider = 'google.firebase.auth';
 /** @hidden */
 export const service = 'firebaseauth.googleapis.com';
 
+/** Resource level options */
+export interface UserOptions {
+  blockingOptions?: {
+    idToken?: boolean;
+    accessToken?: boolean;
+    refreshToken?: boolean;
+  };
+}
+
 /**
  * Handle events related to Firebase authentication users.
  */
-export function user() {
-  return _userWithOptions({});
+export function user(userOptions?: UserOptions) {
+  return _userWithOptions({}, userOptions || {});
 }
 
 /** @hidden */
-export function _userWithOptions(options: DeploymentOptions) {
-  return new UserBuilder(() => {
-    if (!process.env.GCLOUD_PROJECT) {
-      throw new Error('process.env.GCLOUD_PROJECT is not set.');
-    }
-    return 'projects/' + process.env.GCLOUD_PROJECT;
-  }, options);
+export function _userWithOptions(
+  options: DeploymentOptions,
+  userOptions: UserOptions
+) {
+  return new UserBuilder(
+    () => {
+      if (!process.env.GCLOUD_PROJECT) {
+        throw new Error('process.env.GCLOUD_PROJECT is not set.');
+      }
+      return 'projects/' + process.env.GCLOUD_PROJECT;
+    },
+    options,
+    userOptions
+  );
 }
 
 /** Builder used to create Cloud Functions for Firebase Auth user lifecycle events. */
@@ -68,7 +96,8 @@ export class UserBuilder {
   /** @hidden */
   constructor(
     private triggerResource: () => string,
-    private options?: DeploymentOptions
+    private options: DeploymentOptions,
+    private userOptions?: UserOptions
   ) {}
 
   /** Respond to the creation of a Firebase Auth user. */
@@ -83,6 +112,32 @@ export class UserBuilder {
     handler: (user: UserRecord, context: EventContext) => PromiseLike<any> | any
   ): CloudFunction<UserRecord> {
     return this.onOperation(handler, 'user.delete');
+  }
+
+  beforeCreate(
+    handler: (
+      user: AuthUserRecord,
+      context: AuthEventContext
+    ) =>
+      | BeforeCreateResponse
+      | void
+      | Promise<BeforeCreateResponse>
+      | Promise<void>
+  ): BlockingFunction {
+    return this.beforeOperation(handler, 'beforeCreate');
+  }
+
+  beforeSignIn(
+    handler: (
+      user: AuthUserRecord,
+      context: AuthEventContext
+    ) =>
+      | BeforeSignInResponse
+      | void
+      | Promise<BeforeSignInResponse>
+      | Promise<void>
+  ): BlockingFunction {
+    return this.beforeOperation(handler, 'beforeSignIn');
   }
 
   private onOperation(
@@ -102,5 +157,70 @@ export class UserBuilder {
       legacyEventType: `providers/firebase.auth/eventTypes/${eventType}`,
       options: this.options,
     });
+  }
+
+  private beforeOperation(
+    handler: (
+      user: AuthUserRecord,
+      context: AuthEventContext
+    ) =>
+      | BeforeCreateResponse
+      | BeforeSignInResponse
+      | void
+      | Promise<BeforeCreateResponse>
+      | Promise<BeforeSignInResponse>
+      | Promise<void>,
+    eventType: AuthBlockingEventType
+  ): BlockingFunction {
+    const accessToken = this.userOptions?.blockingOptions?.accessToken || false;
+    const idToken = this.userOptions?.blockingOptions?.idToken || false;
+    const refreshToken =
+      this.userOptions?.blockingOptions?.refreshToken || false;
+
+    // Create our own function that just calls the provided function so we know for sure that
+    // handler takes two arguments. This is something common/providers/identity depends on.
+    const wrappedHandler = (user: AuthUserRecord, context: AuthEventContext) =>
+      handler(user, context);
+    const func: any = wrapHandler(eventType, wrappedHandler);
+
+    const legacyEventType = `providers/cloud.auth/eventTypes/user.${eventType}`;
+
+    func.__trigger = {
+      labels: {},
+      ...optionsToTrigger(this.options),
+      blockingTrigger: {
+        eventType: legacyEventType,
+        options: {
+          accessToken,
+          idToken,
+          refreshToken,
+        },
+      },
+    };
+
+    func.__endpoint = {
+      platform: 'gcfv1',
+      labels: {},
+      ...optionsToEndpoint(this.options),
+      blockingTrigger: {
+        eventType: legacyEventType,
+        options: {
+          accessToken,
+          idToken,
+          refreshToken,
+        },
+      },
+    };
+
+    func.__requiredAPIs = [
+      {
+        api: 'identitytoolkit.googleapis.com',
+        reason: 'Needed for auth blocking functions',
+      },
+    ];
+
+    func.run = handler;
+
+    return func;
   }
 }