+++ src/v2/providers/alerts/billing.ts
@@ -1,4 +1,4 @@
-import { getEndpointAnnotation, FirebaseAlertData } from '.';
+import { FirebaseAlertData, getEndpointAnnotation } from '.';
 import { CloudEvent, CloudFunction } from '../../core';
 import * as options from '../../options';
 
@@ -7,9 +7,13 @@ import * as options from '../../options';
  * Payload is wrapped inside a FirebaseAlertData object.
  */
 export interface PlanUpdatePayload {
-  ['@type']: 'com.google.firebase.firebasealerts.PlanUpdatePayload';
+  ['@type']: 'type.googleapis.com/google.events.firebase.firebasealerts.v1.BillingPlanUpdatePayload';
+  /** A Firebase billing plan. */
   billingPlan: string;
+  /** The email address of the person that triggered billing plan change */
   principalEmail: string;
+  /** The type of the notification, e.g. upgrade, downgrade */
+  notificationType: string;
 }
 
 /**
@@ -17,11 +21,15 @@ export interface PlanUpdatePayload {
  * Payload is wrapped inside a FirebaseAlertData object.
  */
 export interface PlanAutomatedUpdatePayload {
-  ['@type']: 'com.google.firebase.firebasealerts.PlanAutomatedUpdatePayload';
+  ['@type']: 'type.googleapis.com/google.events.firebase.firebasealerts.v1.BillingPlanAutomatedUpdatePayload';
+  /** A Firebase billing plan. */
   billingPlan: string;
+  /** The type of the notification, e.g. upgrade, downgrade */
+  notificationType: string;
 }
 
 interface WithAlertType {
+  /** The type of the alerts that got triggered. */
   alertType: string;
 }
 /**