+++ src/v2/providers/storage.ts
@@ -20,11 +20,11 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 // SOFTWARE.
 
-import * as options from '../options';
-import { firebaseConfig } from '../../config';
-import { CloudEvent, CloudFunction } from '../core';
 import { copyIfPresent } from '../../common/encoding';
+import { firebaseConfig } from '../../config';
 import { ManifestEndpoint } from '../../runtime/manifest';
+import { CloudEvent, CloudFunction } from '../core';
+import * as options from '../options';
 
 /**
  * An object within Google Cloud Storage.
@@ -174,6 +174,13 @@ export interface CustomerEncryption {
   keySha256?: string;
 }
 
+interface WithBucket {
+  /** The name of the bucket containing this object. */
+  bucket: string;
+}
+
+export type StorageEvent = CloudEvent<StorageObjectData, WithBucket>;
+
 /** @internal */
 export const archivedEvent = 'google.cloud.storage.object.v1.archived';
 /** @internal */
@@ -191,100 +198,100 @@ export interface StorageOptions extends options.EventHandlerOptions {
 
 /** Handle a storage object archived */
 export function onObjectArchived(
-  handler: (event: CloudEvent<StorageObjectData>) => any | Promise<any>
+  handler: (event: StorageEvent) => any | Promise<any>
 ): CloudFunction<StorageObjectData>;
 
 export function onObjectArchived(
   bucket: string,
-  handler: (event: CloudEvent<StorageObjectData>) => any | Promise<any>
+  handler: (event: StorageEvent) => any | Promise<any>
 ): CloudFunction<StorageObjectData>;
 
 export function onObjectArchived(
   opts: StorageOptions,
-  handler: (event: CloudEvent<StorageObjectData>) => any | Promise<any>
+  handler: (event: StorageEvent) => any | Promise<any>
 ): CloudFunction<StorageObjectData>;
 
 export function onObjectArchived(
   buketOrOptsOrHandler:
     | string
     | StorageOptions
-    | ((event: CloudEvent<StorageObjectData>) => any | Promise<any>),
-  handler?: (event: CloudEvent<StorageObjectData>) => any | Promise<any>
+    | ((event: StorageEvent) => any | Promise<any>),
+  handler?: (event: StorageEvent) => any | Promise<any>
 ): CloudFunction<StorageObjectData> {
   return onOperation(archivedEvent, buketOrOptsOrHandler, handler);
 }
 
 /** Handle a storage object finalized */
 export function onObjectFinalized(
-  handler: (event: CloudEvent<StorageObjectData>) => any | Promise<any>
+  handler: (event: StorageEvent) => any | Promise<any>
 ): CloudFunction<StorageObjectData>;
 
 export function onObjectFinalized(
   bucket: string,
-  handler: (event: CloudEvent<StorageObjectData>) => any | Promise<any>
+  handler: (event: StorageEvent) => any | Promise<any>
 ): CloudFunction<StorageObjectData>;
 
 export function onObjectFinalized(
   opts: StorageOptions,
-  handler: (event: CloudEvent<StorageObjectData>) => any | Promise<any>
+  handler: (event: StorageEvent) => any | Promise<any>
 ): CloudFunction<StorageObjectData>;
 
 export function onObjectFinalized(
   buketOrOptsOrHandler:
     | string
     | StorageOptions
-    | ((event: CloudEvent<StorageObjectData>) => any | Promise<any>),
-  handler?: (event: CloudEvent<StorageObjectData>) => any | Promise<any>
+    | ((event: StorageEvent) => any | Promise<any>),
+  handler?: (event: StorageEvent) => any | Promise<any>
 ): CloudFunction<StorageObjectData> {
   return onOperation(finalizedEvent, buketOrOptsOrHandler, handler);
 }
 
 /** Handle a storage object deleted */
 export function onObjectDeleted(
-  handler: (event: CloudEvent<StorageObjectData>) => any | Promise<any>
+  handler: (event: StorageEvent) => any | Promise<any>
 ): CloudFunction<StorageObjectData>;
 
 export function onObjectDeleted(
   bucket: string,
-  handler: (event: CloudEvent<StorageObjectData>) => any | Promise<any>
+  handler: (event: StorageEvent) => any | Promise<any>
 ): CloudFunction<StorageObjectData>;
 
 export function onObjectDeleted(
   opts: StorageOptions,
-  handler: (event: CloudEvent<StorageObjectData>) => any | Promise<any>
+  handler: (event: StorageEvent) => any | Promise<any>
 ): CloudFunction<StorageObjectData>;
 
 export function onObjectDeleted(
   buketOrOptsOrHandler:
     | string
     | StorageOptions
-    | ((event: CloudEvent<StorageObjectData>) => any | Promise<any>),
-  handler?: (event: CloudEvent<StorageObjectData>) => any | Promise<any>
+    | ((event: StorageEvent) => any | Promise<any>),
+  handler?: (event: StorageEvent) => any | Promise<any>
 ): CloudFunction<StorageObjectData> {
   return onOperation(deletedEvent, buketOrOptsOrHandler, handler);
 }
 
 /** Handle a storage object metadata updated */
 export function onObjectMetadataUpdated(
-  handler: (event: CloudEvent<StorageObjectData>) => any | Promise<any>
+  handler: (event: StorageEvent) => any | Promise<any>
 ): CloudFunction<StorageObjectData>;
 
 export function onObjectMetadataUpdated(
   bucket: string,
-  handler: (event: CloudEvent<StorageObjectData>) => any | Promise<any>
+  handler: (event: StorageEvent) => any | Promise<any>
 ): CloudFunction<StorageObjectData>;
 
 export function onObjectMetadataUpdated(
   opts: StorageOptions,
-  handler: (event: CloudEvent<StorageObjectData>) => any | Promise<any>
+  handler: (event: StorageEvent) => any | Promise<any>
 ): CloudFunction<StorageObjectData>;
 
 export function onObjectMetadataUpdated(
   buketOrOptsOrHandler:
     | string
     | StorageOptions
-    | ((event: CloudEvent<StorageObjectData>) => any | Promise<any>),
-  handler?: (event: CloudEvent<StorageObjectData>) => any | Promise<any>
+    | ((event: StorageEvent) => any | Promise<any>),
+  handler?: (event: StorageEvent) => any | Promise<any>
 ): CloudFunction<StorageObjectData> {
   return onOperation(metadataUpdatedEvent, buketOrOptsOrHandler, handler);
 }
@@ -295,12 +302,12 @@ export function onOperation(
   bucketOrOptsOrHandler:
     | string
     | StorageOptions
-    | ((event: CloudEvent<StorageObjectData>) => any | Promise<any>),
-  handler: (event: CloudEvent<StorageObjectData>) => any | Promise<any>
+    | ((event: StorageEvent) => any | Promise<any>),
+  handler: (event: StorageEvent) => any | Promise<any>
 ): CloudFunction<StorageObjectData> {
   if (typeof bucketOrOptsOrHandler === 'function') {
     handler = bucketOrOptsOrHandler as (
-      event: CloudEvent<StorageObjectData>
+      event: StorageEvent
     ) => any | Promise<any>;
     bucketOrOptsOrHandler = {};
   }
@@ -310,7 +317,7 @@ export function onOperation(
   );
 
   const func = (raw: CloudEvent<unknown>) => {
-    return handler(raw as CloudEvent<StorageObjectData>);
+    return handler(raw as StorageEvent);
   };
 
   func.run = handler;
@@ -382,7 +389,7 @@ export function getOptsAndBucket(
     bucket = bucketOrOpts;
     opts = {};
   } else {
-    bucket = bucketOrOpts.bucket || firebaseConfig().storageBucket;
+    bucket = bucketOrOpts.bucket || firebaseConfig()?.storageBucket;
     opts = { ...bucketOrOpts };
     delete (opts as any).bucket;
   }