+++ CHANGELOG.md
@@ -1,1 +1,2 @@
-- Improve authorization for tasks. (#1073)
+- Adds auth blocking triggers to the auth and identity namespaces (1080).
+- Add support for secrets for v2 triggers (#1079).